파일 데이터 생성

+++ src/main/java/com/ll/exam/spring_fileupload/fileUpload/entity/GenFile.java
@@ -16,7 +16,7 @@ import javax.persistence.Entity;
 @ToString(callSuper = true)
 public class GenFile extends BaseEntity {
     private String relTypeCode;
-    private int relId;
+    private long relId;
     private String typeCode;
     private String type2Code;
     private String fileExtTypeCode;
